add java-doc

+++ src/main/java/com/makarov/factory/generator/manager/EntityDeleter.java
@@ -1,30 +1,39 @@
 package com.makarov.factory.generator.manager;
 
-import com.makarov.factory.generator.expression.adder.ExpressionAdder;
 import com.makarov.annotation.Column;
 import com.makarov.annotation.Table;
+import com.makarov.factory.generator.expression.adder.ExpressionAdder;
 import com.makarov.factory.generator.expression.api.RuleConstruction;
 import com.makarov.factory.generator.expression.impl.RuleExpressions;
 import com.makarov.mapper.util.MapperUtils;
-import com.sun.deploy.util.StringUtils;
 
 import java.lang.reflect.Array;
 import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.List;
 
-
+/**
+ * Class for creating "delete" query by method name
+ *
+ * @author Makarov Alexey
+ * @version 1.0
+ */
 public class EntityDeleter {
-
-    private MapperUtils mapperUtils = new MapperUtils();
 
     private final String queryPattern = "DELETE FROM %s%s%s;";
 
+    /**
+     * Create query by method name
+     *
+     * @param words  - list words in method name
+     * @param params - arguments of method
+     * @return sql-query
+     */
     public String getDeleteQuery(List<String> words, Object[] params) {
         ExpressionAdder adder = new ExpressionAdder(words);
         RuleConstruction rule = new RuleExpressions();
 
-        if (isMappedClass(params[0])) {
+        if (params != null && isMappedClass(params[0])) {
             return deleteByObject(params[0]);
         }
 
@@ -48,8 +57,14 @@ public class EntityDeleter {
         }
     }
 
-    public String deleteByObject(Object entity) {
-        Class<?> clazz = entity.getClass();
+    /**
+     * Create query by method name with parameter as an entity
+     *
+     * @param entity deletable entity
+     * @return sql-query
+     */
+    private String deleteByObject(Object entity) {
+        Class<?> clazz = MapperUtils.getEnhancerSuperclass(entity);
 
         if (!clazz.isAnnotationPresent(Table.class)) {
             return deleteByObjects(entity);
@@ -62,32 +77,38 @@ public class EntityDeleter {
             for (Field field : clazz.getDeclaredFields()) {
                 if (field.isAnnotationPresent(Column.class)) {
                     String columnName = field.getAnnotation(Column.class).name();
-                    String result = columnName + "=" + mapperUtils.getValueField(field, entity);
+                    String result = columnName + "=" + MapperUtils.getValueField(field, entity);
                     expressions.add(result);
                 }
             }
         }
 
         return String.format(queryPattern, table, " WHERE ",
-                StringUtils.join(expressions, " AND "));
+                String.join(" AND ", expressions));
     }
 
-    public String deleteByObjects(Object entity) {
-        Class<?> clazz = entity.getClass();
+    /**
+     * Create query by method name with parameter as an list or array entities
+     *
+     * @param entities deletable entities
+     * @return sql-query
+     */
+    private String deleteByObjects(Object entities) {
+        Class<?> clazz = entities.getClass();
         StringBuilder query = new StringBuilder();
 
         if (clazz.isArray()) {
-            int length = Array.getLength(entity);
+            int length = Array.getLength(entities);
 
             for (int i = 0; i < length; i++) {
-                query.append(deleteByObject(Array.get(entity, i)));
+                query.append(deleteByObject(Array.get(entities, i)));
             }
         } else {
             Iterable iterable;
             try {
-                iterable = Iterable.class.cast(entity);
+                iterable = Iterable.class.cast(entities);
             } catch (ClassCastException exception) {
-                throw new ClassCastException("Saved objects have to be a iterable");
+                throw new ClassCastException("Deleted objects have to be a iterable");
             }
 
             for (Object anIterable : iterable) {
@@ -98,12 +119,23 @@ public class EntityDeleter {
     }
 
 
-    public boolean isMappedClass(Object entity) {
+    /**
+     * Check class for mapping on table
+     *
+     * @param entity - entity class
+     * @return true - class in mapped on table
+     * false - class is not mapped on table
+     */
+    private boolean isMappedClass(Object entity) {
         if (entity == null) {
             return false;
         }
 
         Class<?> clazz = entity.getClass();
+
+        if (MapperUtils.isGeneratedClass(clazz)) {
+            clazz = MapperUtils.getEnhancerSuperclass(entity);
+        }
 
         return clazz.isAnnotationPresent(Table.class)
                 || clazz.isArray()
